pyproject.toml
[tool.poetry]
name = "mqtt-stl"
version = "0.1.0"
description = ""
authors = ["Donald AC Martin <donald@donald-martin.ca>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
paho-mqtt = "^1.6.1"
orjson = "^3.6.6"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = "^22.1"
pylint = "^2.12.2"
typed-ast = "^1.5.2"
pytest-cov = "^3.0.0"
importlib-metadata = "^4.10.1"
typing-extensions = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
